% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC_A}
\alias{BIC_A}
\title{Computes the BIC of a RES distribution with the asymptotic penalty term}
\usage{
BIC_A(S_est, t, mem, rho, psi, eta)
}
\arguments{
\item{S_est}{tensorA::tensor[r, r, ll]. Estimated scatter matrix of cluster ll}

\item{t}{Matrix[N, ll] Squared Mahalanobis distances of data points to cluster centers}

\item{mem}{Matrix[N, ll] Cluster memberships}

\item{rho}{Vector rho of density generator}

\item{psi}{Vector. psi of density generator}

\item{eta}{Vector. eta of density generator}
}
\value{
list
\enumerate{
\item bic
\item pen : Penalty term
\item like : Likelihood term
}
}
\description{
Computes the BIC of a RES distribution with the asymptotic penalty term
}
\examples{


}
