% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC_F}
\alias{BIC_F}
\title{computes the BIC of a RES distribution based on a finite sample penalty term}
\usage{
BIC_F(data, S_est, mu_est, t, mem, rho, psi, eta)
}
\arguments{
\item{data}{Matrix[N, r] data samples}

\item{S_est}{tensorA::Tensor[r, r, ll] Scatter matrices of the clusters}

\item{mu_est}{Matrix[r, ll] estimated mean vectors of all clusters}

\item{t}{Matrix[N, ll] Squared Mahalanobis distances to cluster centers}

\item{mem}{Matrix[N, ll] cluster memberships}

\item{rho}{Vector rho of density generator}

\item{psi}{Vector. psi of density generator}

\item{eta}{Vector. eta of density generator}
}
\value{
list
\enumerate{
\item bic
\item pen : Penalty term
\item like : Likelihood term
}
}
\description{
computes the BIC of a RES distribution based on a finite sample penalty term
}
\note{
"Robust M-Estimation based Bayesian Cluster Enumeration for Real Elliptically Symmetric Distributions"
Christian A. Schroth and Michael Muma, Signal Processing Group, Technische Universität Darmstadt
submitted to IEEE Transactions on Signal Processing
}
