% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicationMatrix.R
\name{duplicationMatrix}
\alias{duplicationMatrix}
\title{The duplication matrix}
\usage{
duplicationMatrix(num_features)
}
\arguments{
\item{num_features}{int. Number of features}
}
\value{
dupmat Matrix[num_features^2, .5*num_features*(num_features+1)] The duplication matrix
}
\description{
The duplication matrix
}
\note{
[1] F. K. Teklehaymanot, M. Muma, and A. M. Zoubir, "Bayesian Cluster Enumeration Criterion for Unsupervised Learning",
IEEE Trans. Signal Process. (accepted),
[Online-Edition: https://arxiv.org/abs/1710.07954v2], 2018.

[2] F. K. Teklehaymanot, M. Muma, and A. M. Zoubir, "Novel Bayesian Cluster
    Enumeration Criterion for Cluster Analysis With Finite Sample Penalty Term",
in Proc. 43rd IEEE Int. conf. on Acoustics, Speech and Signal Process. (ICASSP), pp. 4274-4278, 2018,
[Online-Edition: https://www.researchgate.net/publication/322918028]


Copyright (c) 2018 Freweyni K. Teklehaymanot. All rights reserved.
}
