% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_RES.R
\name{EM_RES}
\alias{EM_RES}
\title{Title}
\usage{
EM_RES(data, ll, g, psi, limit = 1e-06, em_max_iter = 200,
  reg_value = 1e-06, test_args = NULL)
}
\arguments{
\item{data}{Matrix[N, r] data without labels}

\item{ll}{scalar, number of clusters}

\item{g}{function(t), gauss}

\item{reg_value}{}
}
\description{
Title
}
