% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{eta_gaus}
\alias{eta_gaus}
\title{Computes eta(t) of the Gaussian distribution}
\usage{
eta_gaus(t)
}
\arguments{
\item{t}{vector of length N. Squared Mahalanobis distances}
}
\value{
vector of length N. eta of a Gaussian distribution
}
\description{
Computes eta(t) of the Gaussian distribution
}
\examples{

eta_gaus(c(1,2,3))

}
