% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{eta_huber}
\alias{eta_huber}
\title{Computes eta(t) of the Huber distribution}
\usage{
eta_huber(t, r, lst = list())
}
\arguments{
\item{t}{Vector. Squared Mahalanobis distances}

\item{r}{Scalar. Dimension}

\item{lst}{List. Either empty, Contains field qH or the fields cH and bH. Default is set to empty}
}
\value{
eta of the Huber distribution
}
\description{
Computes eta(t) of the Huber distribution
}
\examples{
eta_huber(rnorm(4), 2, list(.1))
eta_huber(rnorm(4), 2)

}
