% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{eta_tukey}
\alias{eta_tukey}
\title{Computes eta(t) of the Tukey loss function}
\usage{
eta_tukey(t, cT)
}
\arguments{
\item{t}{Vector. Squared Mahalanobis distances}

\item{cT}{Scalar tuning parameter}
}
\value{
eta of Tukey
}
\description{
Computes eta(t) of the Tukey loss function
}
\examples{

eta_tukey(c(-1.4456554,  0.9074437,  0.2274346, -0.8734823), .1)

}
