% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM_RES.R
\name{FIM_RES}
\alias{FIM_RES}
\title{computes FIM of one cluster for a given RES distribution}
\usage{
FIM_RES(x_hat, t_m, S_est, psi, eta, D)
}
\arguments{
\item{x_hat}{Matrix[r, N_m] data matrix}

\item{t_m}{vector[N_m] squared Mahalanobis distances}

\item{S_est}{Matrix[r, r] Scatter matrix}

\item{psi}{scalar, psi of density generator}

\item{eta}{scalar, eta of density generator}

\item{D}{Matrix[r^2, 1/2*r*(r+1)] duplication matrix}
}
\value{
J Matrix[q, q] FIM
}
\description{
computes FIM of one cluster for a given RES distribution
}
