% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matSum}
\alias{matSum}
\title{sum of matrices of a rank 3 tensor}
\usage{
matSum(tensor)
}
\arguments{
\item{tensor}{3darray}
}
\value{
Sum of matrices in tensor
}
\description{
sum of matrices of a rank 3 tensor
}
