% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestSPD.R
\name{nearestSPD}
\alias{nearestSPD}
\title{Finds closest symmetric positive definite matrix}
\usage{
nearestSPD(A)
}
\arguments{
\item{A}{matrix}
}
\value{
Ahat closest SPD matrix to A
}
\description{
Finds closest symmetric positive definite matrix
}
\examples{

m <- matrix(c(1,2,3,4,5,6,7,8,9),3,3,byrow = TRUE)
nearestSPD(m)
}
