% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho_gaus}
\alias{rho_gaus}
\title{rho(t) of the Gaussian distributiomn}
\usage{
rho_gaus(t, r)
}
\arguments{
\item{t}{vector[N] squared Mahalanobis distances}

\item{r}{scaler, dimension}
}
\value{
rho rho of the Gaussian
}
\description{
rho(t) of the Gaussian distributiomn
}
\note{
"Robust M-Estimation based Bayesian Cluster Enumeration for Real Elliptically Symmetric Distributions"
Christian A. Schroth and Michael Muma, Signal Processing Group, Technische Universität Darmstadt
submitted to IEEE Transactions on Signal Processing
}
\examples{
rho_gaus(rnorm(5), 1)
}
