% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho_huber}
\alias{rho_huber}
\title{Computes rho(t) of the Huber distribution}
\usage{
rho_huber(t, r, lst = list())
}
\arguments{
\item{t}{vector[N] squared Mahalanobis distances}

\item{r}{scalar, dimension}

\item{lst}{List of scalar tuning parameters. There are 3 valid options of this argument
\enumerate{
\item Empty list. Default case
\item List containing qH
\item List containing cH, bH, aH
}}
}
\value{
vector[N] rho(t) of Huber distribution
}
\description{
Computes rho(t) of the Huber distribution
}
\examples{
rho_huber(rnorm(5), 2)
}
