% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho_t}
\alias{rho_t}
\title{Computes rho(t) of the t distribution}
\usage{
rho_t(t, r, nu)
}
\arguments{
\item{t}{vector[N] squared Mahalanobis distance}

\item{r}{scalar, dimension}

\item{nu}{scalar, degree of freedom}
}
\value{
vector[N] rho(t) of the t distribution
}
\description{
Computes rho(t) of the t distribution
}
\examples{
rho_t(rnorm(5),2,5)
}
