% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho_tukey}
\alias{rho_tukey}
\title{Computes rho(t) of the Tukey loss function}
\usage{
rho_tukey(t, r, cT)
}
\arguments{
\item{t}{vector[N] squared Mahalanobis distance}

\item{r}{scalar, dimension}

\item{cT}{scalar, tuning parameter}
}
\value{
rho vector[N] rho of the Tukey loss function
}
\description{
Computes rho(t) of the Tukey loss function
}
\note{
"Robust M-Estimation based Bayesian Cluster Enumeration for Real Elliptically Symmetric Distributions"
Christian A. Schroth and Michael Muma, Signal Processing Group, Technische Universität Darmstadt
submitted to IEEE Transactions on Signal Processing
}
\examples{

rho_tukey(rnorm(5),2,5)
}
