% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{eta_t}
\alias{eta_t}
\title{Computes eta(t) of the t distribution}
\usage{
eta_t(t, r, nu)
}
\arguments{
\item{t}{Vector. Squared Mahalanobis distances}

\item{r}{int. Dimension}

\item{nu}{int. Degree of Freedom}
}
\value{
eta(t) of the t distribution
}
\description{
Computes eta(t) of the t distribution
}
\examples{
eta_t(c(-0.29022179, -0.58196155, -0.51125062, -0.09460791), .1, .3)


}
